third_party/blink/web_tests/fast/dom/domstring-attribute-reflection-expected.txt

Reflected DOMString attribute test for button/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for fieldset/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for form/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for input/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for input/@step
Initial value:
PASS element.step is ""
PASS element.getAttribute("step") is null
Setting a value via the IDL attribute:
PASS element.step = "foo"; element.step is "foo"
PASS element.getAttribute("step") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("step", " bar\n"); element.step is " bar\n"
PASS element.getAttribute("step") is " bar\n"
Setting null via the IDL attribute:
PASS element.step = null; element.step is "null"
PASS element.getAttribute("step") is "null"
Setting null via the content attribute:
PASS element.setAttribute("step", null); element.step is "null"
PASS element.getAttribute("step") is "null"
Setting undefined via the IDL attribute:
PASS element.step = undefined; element.step is "undefined"
PASS element.getAttribute("step") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("step", undefined); element.step is "undefined"
PASS element.getAttribute("step") is "undefined"
Setting non-string via the IDL attribute:
PASS element.step = 123; element.step is "123"
PASS element.getAttribute("step") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("step", 456); element.step is "456"
PASS element.getAttribute("step") is "456"


Reflected DOMString attribute test for object/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for output/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for select/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for textarea/@name
Initial value:
PASS element.name is ""
PASS element.getAttribute("name") is null
Setting a value via the IDL attribute:
PASS element.name = "foo"; element.name is "foo"
PASS element.getAttribute("name") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("name", " bar\n"); element.name is " bar\n"
PASS element.getAttribute("name") is " bar\n"
Setting null via the IDL attribute:
PASS element.name = null; element.name is "null"
PASS element.getAttribute("name") is "null"
Setting null via the content attribute:
PASS element.setAttribute("name", null); element.name is "null"
PASS element.getAttribute("name") is "null"
Setting undefined via the IDL attribute:
PASS element.name = undefined; element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("name", undefined); element.name is "undefined"
PASS element.getAttribute("name") is "undefined"
Setting non-string via the IDL attribute:
PASS element.name = 123; element.name is "123"
PASS element.getAttribute("name") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("name", 456); element.name is "456"
PASS element.getAttribute("name") is "456"


Reflected DOMString attribute test for frame/@marginheight
Initial value:
PASS element.marginHeight is ""
PASS element.getAttribute("marginheight") is null
Setting a value via the IDL attribute:
PASS element.marginHeight = "foo"; element.marginHeight is "foo"
PASS element.getAttribute("marginheight") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("marginheight", " bar\n"); element.marginHeight is " bar\n"
PASS element.getAttribute("marginheight") is " bar\n"
Setting null via the IDL attribute:
PASS element.marginHeight = null; element.marginHeight is ""
PASS element.getAttribute("marginheight") is ""
Setting null via the content attribute:
PASS element.setAttribute("marginheight", null); element.marginHeight is "null"
PASS element.getAttribute("marginheight") is "null"
Setting undefined via the IDL attribute:
PASS element.marginHeight = undefined; element.marginHeight is "undefined"
PASS element.getAttribute("marginheight") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("marginheight", undefined); element.marginHeight is "undefined"
PASS element.getAttribute("marginheight") is "undefined"
Setting non-string via the IDL attribute:
PASS element.marginHeight = 123; element.marginHeight is "123"
PASS element.getAttribute("marginheight") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("marginheight", 456); element.marginHeight is "456"
PASS element.getAttribute("marginheight") is "456"


Reflected DOMString attribute test for frame/@marginwidth
Initial value:
PASS element.marginWidth is ""
PASS element.getAttribute("marginwidth") is null
Setting a value via the IDL attribute:
PASS element.marginWidth = "foo"; element.marginWidth is "foo"
PASS element.getAttribute("marginwidth") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("marginwidth", " bar\n"); element.marginWidth is " bar\n"
PASS element.getAttribute("marginwidth") is " bar\n"
Setting null via the IDL attribute:
PASS element.marginWidth = null; element.marginWidth is ""
PASS element.getAttribute("marginwidth") is ""
Setting null via the content attribute:
PASS element.setAttribute("marginwidth", null); element.marginWidth is "null"
PASS element.getAttribute("marginwidth") is "null"
Setting undefined via the IDL attribute:
PASS element.marginWidth = undefined; element.marginWidth is "undefined"
PASS element.getAttribute("marginwidth") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("marginwidth", undefined); element.marginWidth is "undefined"
PASS element.getAttribute("marginwidth") is "undefined"
Setting non-string via the IDL attribute:
PASS element.marginWidth = 123; element.marginWidth is "123"
PASS element.getAttribute("marginwidth") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("marginwidth", 456); element.marginWidth is "456"
PASS element.getAttribute("marginwidth") is "456"


Reflected DOMString attribute test for iframe/@marginheight
Initial value:
PASS element.marginHeight is ""
PASS element.getAttribute("marginheight") is null
Setting a value via the IDL attribute:
PASS element.marginHeight = "foo"; element.marginHeight is "foo"
PASS element.getAttribute("marginheight") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("marginheight", " bar\n"); element.marginHeight is " bar\n"
PASS element.getAttribute("marginheight") is " bar\n"
Setting null via the IDL attribute:
PASS element.marginHeight = null; element.marginHeight is ""
PASS element.getAttribute("marginheight") is ""
Setting null via the content attribute:
PASS element.setAttribute("marginheight", null); element.marginHeight is "null"
PASS element.getAttribute("marginheight") is "null"
Setting undefined via the IDL attribute:
PASS element.marginHeight = undefined; element.marginHeight is "undefined"
PASS element.getAttribute("marginheight") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("marginheight", undefined); element.marginHeight is "undefined"
PASS element.getAttribute("marginheight") is "undefined"
Setting non-string via the IDL attribute:
PASS element.marginHeight = 123; element.marginHeight is "123"
PASS element.getAttribute("marginheight") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("marginheight", 456); element.marginHeight is "456"
PASS element.getAttribute("marginheight") is "456"


Reflected DOMString attribute test for iframe/@marginwidth
Initial value:
PASS element.marginWidth is ""
PASS element.getAttribute("marginwidth") is null
Setting a value via the IDL attribute:
PASS element.marginWidth = "foo"; element.marginWidth is "foo"
PASS element.getAttribute("marginwidth") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("marginwidth", " bar\n"); element.marginWidth is " bar\n"
PASS element.getAttribute("marginwidth") is " bar\n"
Setting null via the IDL attribute:
PASS element.marginWidth = null; element.marginWidth is ""
PASS element.getAttribute("marginwidth") is ""
Setting null via the content attribute:
PASS element.setAttribute("marginwidth", null); element.marginWidth is "null"
PASS element.getAttribute("marginwidth") is "null"
Setting undefined via the IDL attribute:
PASS element.marginWidth = undefined; element.marginWidth is "undefined"
PASS element.getAttribute("marginwidth") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("marginwidth", undefined); element.marginWidth is "undefined"
PASS element.getAttribute("marginwidth") is "undefined"
Setting non-string via the IDL attribute:
PASS element.marginWidth = 123; element.marginWidth is "123"
PASS element.getAttribute("marginwidth") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("marginwidth", 456); element.marginWidth is "456"
PASS element.getAttribute("marginwidth") is "456"


Reflected DOMString attribute test for body/@text
Initial value:
PASS element.text is ""
PASS element.getAttribute("text") is null
Setting a value via the IDL attribute:
PASS element.text = "foo"; element.text is "foo"
PASS element.getAttribute("text") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("text", " bar\n"); element.text is " bar\n"
PASS element.getAttribute("text") is " bar\n"
Setting null via the IDL attribute:
PASS element.text = null; element.text is ""
PASS element.getAttribute("text") is ""
Setting null via the content attribute:
PASS element.setAttribute("text", null); element.text is "null"
PASS element.getAttribute("text") is "null"
Setting undefined via the IDL attribute:
PASS element.text = undefined; element.text is "undefined"
PASS element.getAttribute("text") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("text", undefined); element.text is "undefined"
PASS element.getAttribute("text") is "undefined"
Setting non-string via the IDL attribute:
PASS element.text = 123; element.text is "123"
PASS element.getAttribute("text") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("text", 456); element.text is "456"
PASS element.getAttribute("text") is "456"


Reflected DOMString attribute test for body/@link
Initial value:
PASS element.link is ""
PASS element.getAttribute("link") is null
Setting a value via the IDL attribute:
PASS element.link = "foo"; element.link is "foo"
PASS element.getAttribute("link") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("link", " bar\n"); element.link is " bar\n"
PASS element.getAttribute("link") is " bar\n"
Setting null via the IDL attribute:
PASS element.link = null; element.link is ""
PASS element.getAttribute("link") is ""
Setting null via the content attribute:
PASS element.setAttribute("link", null); element.link is "null"
PASS element.getAttribute("link") is "null"
Setting undefined via the IDL attribute:
PASS element.link = undefined; element.link is "undefined"
PASS element.getAttribute("link") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("link", undefined); element.link is "undefined"
PASS element.getAttribute("link") is "undefined"
Setting non-string via the IDL attribute:
PASS element.link = 123; element.link is "123"
PASS element.getAttribute("link") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("link", 456); element.link is "456"
PASS element.getAttribute("link") is "456"


Reflected DOMString attribute test for body/@alink
Initial value:
PASS element.aLink is ""
PASS element.getAttribute("alink") is null
Setting a value via the IDL attribute:
PASS element.aLink = "foo"; element.aLink is "foo"
PASS element.getAttribute("alink") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("alink", " bar\n"); element.aLink is " bar\n"
PASS element.getAttribute("alink") is " bar\n"
Setting null via the IDL attribute:
PASS element.aLink = null; element.aLink is ""
PASS element.getAttribute("alink") is ""
Setting null via the content attribute:
PASS element.setAttribute("alink", null); element.aLink is "null"
PASS element.getAttribute("alink") is "null"
Setting undefined via the IDL attribute:
PASS element.aLink = undefined; element.aLink is "undefined"
PASS element.getAttribute("alink") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("alink", undefined); element.aLink is "undefined"
PASS element.getAttribute("alink") is "undefined"
Setting non-string via the IDL attribute:
PASS element.aLink = 123; element.aLink is "123"
PASS element.getAttribute("alink") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("alink", 456); element.aLink is "456"
PASS element.getAttribute("alink") is "456"


Reflected DOMString attribute test for body/@vlink
Initial value:
PASS element.vLink is ""
PASS element.getAttribute("vlink") is null
Setting a value via the IDL attribute:
PASS element.vLink = "foo"; element.vLink is "foo"
PASS element.getAttribute("vlink") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("vlink", " bar\n"); element.vLink is " bar\n"
PASS element.getAttribute("vlink") is " bar\n"
Setting null via the IDL attribute:
PASS element.vLink = null; element.vLink is ""
PASS element.getAttribute("vlink") is ""
Setting null via the content attribute:
PASS element.setAttribute("vlink", null); element.vLink is "null"
PASS element.getAttribute("vlink") is "null"
Setting undefined via the IDL attribute:
PASS element.vLink = undefined; element.vLink is "undefined"
PASS element.getAttribute("vlink") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("vlink", undefined); element.vLink is "undefined"
PASS element.getAttribute("vlink") is "undefined"
Setting non-string via the IDL attribute:
PASS element.vLink = 123; element.vLink is "123"
PASS element.getAttribute("vlink") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("vlink", 456); element.vLink is "456"
PASS element.getAttribute("vlink") is "456"


Reflected DOMString attribute test for body/@bgcolor
Initial value:
PASS element.bgColor is ""
PASS element.getAttribute("bgcolor") is null
Setting a value via the IDL attribute:
PASS element.bgColor = "foo"; element.bgColor is "foo"
PASS element.getAttribute("bgcolor") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("bgcolor", " bar\n"); element.bgColor is " bar\n"
PASS element.getAttribute("bgcolor") is " bar\n"
Setting null via the IDL attribute:
PASS element.bgColor = null; element.bgColor is ""
PASS element.getAttribute("bgcolor") is ""
Setting null via the content attribute:
PASS element.setAttribute("bgcolor", null); element.bgColor is "null"
PASS element.getAttribute("bgcolor") is "null"
Setting undefined via the IDL attribute:
PASS element.bgColor = undefined; element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("bgcolor", undefined); element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting non-string via the IDL attribute:
PASS element.bgColor = 123; element.bgColor is "123"
PASS element.getAttribute("bgcolor") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("bgcolor", 456); element.bgColor is "456"
PASS element.getAttribute("bgcolor") is "456"


Reflected DOMString attribute test for font/@color
Initial value:
PASS element.color is ""
PASS element.getAttribute("color") is null
Setting a value via the IDL attribute:
PASS element.color = "foo"; element.color is "foo"
PASS element.getAttribute("color") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("color", " bar\n"); element.color is " bar\n"
PASS element.getAttribute("color") is " bar\n"
Setting null via the IDL attribute:
PASS element.color = null; element.color is ""
PASS element.getAttribute("color") is ""
Setting null via the content attribute:
PASS element.setAttribute("color", null); element.color is "null"
PASS element.getAttribute("color") is "null"
Setting undefined via the IDL attribute:
PASS element.color = undefined; element.color is "undefined"
PASS element.getAttribute("color") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("color", undefined); element.color is "undefined"
PASS element.getAttribute("color") is "undefined"
Setting non-string via the IDL attribute:
PASS element.color = 123; element.color is "123"
PASS element.getAttribute("color") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("color", 456); element.color is "456"
PASS element.getAttribute("color") is "456"


Reflected DOMString attribute test for img/@border
Initial value:
PASS element.border is ""
PASS element.getAttribute("border") is null
Setting a value via the IDL attribute:
PASS element.border = "foo"; element.border is "foo"
PASS element.getAttribute("border") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("border", " bar\n"); element.border is " bar\n"
PASS element.getAttribute("border") is " bar\n"
Setting null via the IDL attribute:
PASS element.border = null; element.border is ""
PASS element.getAttribute("border") is ""
Setting null via the content attribute:
PASS element.setAttribute("border", null); element.border is "null"
PASS element.getAttribute("border") is "null"
Setting undefined via the IDL attribute:
PASS element.border = undefined; element.border is "undefined"
PASS element.getAttribute("border") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("border", undefined); element.border is "undefined"
PASS element.getAttribute("border") is "undefined"
Setting non-string via the IDL attribute:
PASS element.border = 123; element.border is "123"
PASS element.getAttribute("border") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("border", 456); element.border is "456"
PASS element.getAttribute("border") is "456"


Reflected DOMString attribute test for object/@border
Initial value:
PASS element.border is ""
PASS element.getAttribute("border") is null
Setting a value via the IDL attribute:
PASS element.border = "foo"; element.border is "foo"
PASS element.getAttribute("border") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("border", " bar\n"); element.border is " bar\n"
PASS element.getAttribute("border") is " bar\n"
Setting null via the IDL attribute:
PASS element.border = null; element.border is ""
PASS element.getAttribute("border") is ""
Setting null via the content attribute:
PASS element.setAttribute("border", null); element.border is "null"
PASS element.getAttribute("border") is "null"
Setting undefined via the IDL attribute:
PASS element.border = undefined; element.border is "undefined"
PASS element.getAttribute("border") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("border", undefined); element.border is "undefined"
PASS element.getAttribute("border") is "undefined"
Setting non-string via the IDL attribute:
PASS element.border = 123; element.border is "123"
PASS element.getAttribute("border") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("border", 456); element.border is "456"
PASS element.getAttribute("border") is "456"


Reflected DOMString attribute test for table/@bgcolor
Initial value:
PASS element.bgColor is ""
PASS element.getAttribute("bgcolor") is null
Setting a value via the IDL attribute:
PASS element.bgColor = "foo"; element.bgColor is "foo"
PASS element.getAttribute("bgcolor") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("bgcolor", " bar\n"); element.bgColor is " bar\n"
PASS element.getAttribute("bgcolor") is " bar\n"
Setting null via the IDL attribute:
PASS element.bgColor = null; element.bgColor is ""
PASS element.getAttribute("bgcolor") is ""
Setting null via the content attribute:
PASS element.setAttribute("bgcolor", null); element.bgColor is "null"
PASS element.getAttribute("bgcolor") is "null"
Setting undefined via the IDL attribute:
PASS element.bgColor = undefined; element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("bgcolor", undefined); element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting non-string via the IDL attribute:
PASS element.bgColor = 123; element.bgColor is "123"
PASS element.getAttribute("bgcolor") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("bgcolor", 456); element.bgColor is "456"
PASS element.getAttribute("bgcolor") is "456"


Reflected DOMString attribute test for table/@cellpadding
Initial value:
PASS element.cellPadding is ""
PASS element.getAttribute("cellpadding") is null
Setting a value via the IDL attribute:
PASS element.cellPadding = "foo"; element.cellPadding is "foo"
PASS element.getAttribute("cellpadding") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("cellpadding", " bar\n"); element.cellPadding is " bar\n"
PASS element.getAttribute("cellpadding") is " bar\n"
Setting null via the IDL attribute:
PASS element.cellPadding = null; element.cellPadding is ""
PASS element.getAttribute("cellpadding") is ""
Setting null via the content attribute:
PASS element.setAttribute("cellpadding", null); element.cellPadding is "null"
PASS element.getAttribute("cellpadding") is "null"
Setting undefined via the IDL attribute:
PASS element.cellPadding = undefined; element.cellPadding is "undefined"
PASS element.getAttribute("cellpadding") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("cellpadding", undefined); element.cellPadding is "undefined"
PASS element.getAttribute("cellpadding") is "undefined"
Setting non-string via the IDL attribute:
PASS element.cellPadding = 123; element.cellPadding is "123"
PASS element.getAttribute("cellpadding") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("cellpadding", 456); element.cellPadding is "456"
PASS element.getAttribute("cellpadding") is "456"


Reflected DOMString attribute test for table/@cellspacing
Initial value:
PASS element.cellSpacing is ""
PASS element.getAttribute("cellspacing") is null
Setting a value via the IDL attribute:
PASS element.cellSpacing = "foo"; element.cellSpacing is "foo"
PASS element.getAttribute("cellspacing") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("cellspacing", " bar\n"); element.cellSpacing is " bar\n"
PASS element.getAttribute("cellspacing") is " bar\n"
Setting null via the IDL attribute:
PASS element.cellSpacing = null; element.cellSpacing is ""
PASS element.getAttribute("cellspacing") is ""
Setting null via the content attribute:
PASS element.setAttribute("cellspacing", null); element.cellSpacing is "null"
PASS element.getAttribute("cellspacing") is "null"
Setting undefined via the IDL attribute:
PASS element.cellSpacing = undefined; element.cellSpacing is "undefined"
PASS element.getAttribute("cellspacing") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("cellspacing", undefined); element.cellSpacing is "undefined"
PASS element.getAttribute("cellspacing") is "undefined"
Setting non-string via the IDL attribute:
PASS element.cellSpacing = 123; element.cellSpacing is "123"
PASS element.getAttribute("cellspacing") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("cellspacing", 456); element.cellSpacing is "456"
PASS element.getAttribute("cellspacing") is "456"


Reflected DOMString attribute test for td/@bgcolor
Initial value:
PASS element.bgColor is ""
PASS element.getAttribute("bgcolor") is null
Setting a value via the IDL attribute:
PASS element.bgColor = "foo"; element.bgColor is "foo"
PASS element.getAttribute("bgcolor") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("bgcolor", " bar\n"); element.bgColor is " bar\n"
PASS element.getAttribute("bgcolor") is " bar\n"
Setting null via the IDL attribute:
PASS element.bgColor = null; element.bgColor is ""
PASS element.getAttribute("bgcolor") is ""
Setting null via the content attribute:
PASS element.setAttribute("bgcolor", null); element.bgColor is "null"
PASS element.getAttribute("bgcolor") is "null"
Setting undefined via the IDL attribute:
PASS element.bgColor = undefined; element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("bgcolor", undefined); element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting non-string via the IDL attribute:
PASS element.bgColor = 123; element.bgColor is "123"
PASS element.getAttribute("bgcolor") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("bgcolor", 456); element.bgColor is "456"
PASS element.getAttribute("bgcolor") is "456"


Reflected DOMString attribute test for th/@bgcolor
Initial value:
PASS element.bgColor is ""
PASS element.getAttribute("bgcolor") is null
Setting a value via the IDL attribute:
PASS element.bgColor = "foo"; element.bgColor is "foo"
PASS element.getAttribute("bgcolor") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("bgcolor", " bar\n"); element.bgColor is " bar\n"
PASS element.getAttribute("bgcolor") is " bar\n"
Setting null via the IDL attribute:
PASS element.bgColor = null; element.bgColor is ""
PASS element.getAttribute("bgcolor") is ""
Setting null via the content attribute:
PASS element.setAttribute("bgcolor", null); element.bgColor is "null"
PASS element.getAttribute("bgcolor") is "null"
Setting undefined via the IDL attribute:
PASS element.bgColor = undefined; element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("bgcolor", undefined); element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting non-string via the IDL attribute:
PASS element.bgColor = 123; element.bgColor is "123"
PASS element.getAttribute("bgcolor") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("bgcolor", 456); element.bgColor is "456"
PASS element.getAttribute("bgcolor") is "456"


Reflected DOMString attribute test for tr/@bgcolor
Initial value:
PASS element.bgColor is ""
PASS element.getAttribute("bgcolor") is null
Setting a value via the IDL attribute:
PASS element.bgColor = "foo"; element.bgColor is "foo"
PASS element.getAttribute("bgcolor") is "foo"
Setting a value via the content attribute:
PASS element.setAttribute("bgcolor", " bar\n"); element.bgColor is " bar\n"
PASS element.getAttribute("bgcolor") is " bar\n"
Setting null via the IDL attribute:
PASS element.bgColor = null; element.bgColor is ""
PASS element.getAttribute("bgcolor") is ""
Setting null via the content attribute:
PASS element.setAttribute("bgcolor", null); element.bgColor is "null"
PASS element.getAttribute("bgcolor") is "null"
Setting undefined via the IDL attribute:
PASS element.bgColor = undefined; element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting undefined via the content attribute:
PASS element.setAttribute("bgcolor", undefined); element.bgColor is "undefined"
PASS element.getAttribute("bgcolor") is "undefined"
Setting non-string via the IDL attribute:
PASS element.bgColor = 123; element.bgColor is "123"
PASS element.getAttribute("bgcolor") is "123"
Setting non-string via the content attribute:
PASS element.setAttribute("bgcolor", 456); element.bgColor is "456"
PASS element.getAttribute("bgcolor") is "456"


PASS successfullyParsed is true

TEST COMPLETE
